Add auth check

+++ app/routes/capture.js
@@ -2,14 +2,16 @@ const ViewModel = require('./models/search')
 const { getDebts, deleteDebt } = require('../debt')
 const schema = require('./schemas/capture')
 const Joi = require('joi')
+const { enrichment } = require('../auth/permissions')
+const ensureHasPermission = require('../ensure-has-permission')
 const searchLabelText = 'Search for data by FRN number'
 
 module.exports = [{
   method: 'GET',
   path: '/capture',
   options: {
-    auth: false,
     handler: async (request, h) => {
+      await ensureHasPermission(request, h, [enrichment])
       const captureData = await getDebts()
       return h.view('capture', { captureData, ...new ViewModel(searchLabelText) })
     }
@@ -19,15 +21,16 @@ module.exports = [{
   method: 'POST',
   path: '/capture',
   options: {
-    auth: false,
     validate: {
       payload: schema,
       failAction: async (request, h, error) => {
+        await ensureHasPermission(request, h, [enrichment])
         const captureData = await getDebts()
         return h.view('capture', { captureData, ...new ViewModel(searchLabelText, request.payload.frn, error) }).code(400).takeover()
       }
     },
     handler: async (request, h) => {
+      await ensureHasPermission(request, h, [enrichment])
       const frn = request.payload.frn
       const captureData = await getDebts()
       const filteredCaptureData = captureData.filter(x => x.frn === String(frn))
@@ -48,11 +51,13 @@ module.exports = [{
         debtDataId: Joi.number().integer().required()
       }),
       failAction: async (request, h, error) => {
+        await ensureHasPermission(request, h, [enrichment])
         const captureData = await getDebts()
         return h.view('capture', { captureData, ...new ViewModel(searchLabelText, request.payload.frn, error) }).code(400).takeover()
       }
     },
     handler: async (request, h) => {
+      await ensureHasPermission(request, h, [enrichment])
       await deleteDebt(request.payload.debtDataId)
       return h.redirect('/capture')
     }
